requirements.txt
argparse==1.4.0
progress==1.5
mysql-connector==2.2.9
